+++ src/Achievement.php
@@ -14,7 +14,7 @@ class Achievement
      *
      * @param array $data
      */
-    public function __construct(array $data = [])
+    public function __construct(array $data)
     {
         $this->data = $data;
     }