+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.9.4] - 2017-07-02
+### Changed
+- `AchievementsManager::updateAchievementCriteria` renamed to `updateAchievementCriterias`;
+- `array $data = []` argument from `AchievementsManager::updateAchievementCriteria` changed to `$data = null`;
+- `Achievement` constructor signature changed from `(array $data = [])` to `(array $data`);
+- `AchievementCriteriaChange` now uses `AchievementCriteriaChange::PROGRESS_ACCUMULATE` as default progress type (instead of `PROGRESS_HIGHEST`);
+
 ## [0.9.3] - 2017-07-02
 ### Added
 - Optional `array $progressData` argument to `AchievementCriteriaChange` constructor;
@@ -28,7 +35,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.9.0] - 2017-06-30
 ### Initial release.
 
-[Unreleased]: https://github.com/tzurbaev/achievements/compare/0.9.3...HEAD
+[Unreleased]: https://github.com/tzurbaev/achievements/compare/0.9.4...HEAD
+[0.9.4]: https://github.com/tzurbaev/achievements/compare/0.9.1...0.9.4
 [0.9.3]: https://github.com/tzurbaev/achievements/compare/0.9.1...0.9.3
 [0.9.2]: https://github.com/tzurbaev/achievements/compare/0.9.1...0.9.2
 [0.9.1]: https://github.com/tzurbaev/achievements/compare/0.9.0...0.9.1